[Add]Union正式実装

+++ OpenSiv3D_0.6.152_Test/UIManager.h
@@ -30,7 +30,7 @@ public:
 	void UIResultInit();
 	void UIResultUpdate();
 	bool ChackTime();
-	void AddScoreRef(float score);
+	void AddScoreRef(int score);
 	void ResetUI();
 	float CurrentTime();
 	int32 Score();
